requirements.txt
psycopg2-binary
alembic
flask